feature: split grpc methods implementation from the server methods
